+++ eslint.config.js
@@ -1,8 +1,10 @@
-import markdown from 'eslint-plugin-markdown';
-import globals from 'globals';
-import n from 'eslint-plugin-n';
+import markdown from '@eslint/markdown';
+import { defineConfig } from 'eslint/config';
 import svelteOrgEslintConfig from '@sveltejs/eslint-config';
-export default [
+import n from 'eslint-plugin-n';
+import globals from 'globals';
+
+export default defineConfig([
 	{
 		name: 'local/ignores',
 		ignores: [
@@ -32,7 +34,14 @@ export default [
 		}
 	},
 	n.configs['flat/recommended-module'],
-	...markdown.configs.recommended,
+	{
+		name: 'local/markdown',
+		files: ['**/*.md'],
+		plugins: {
+			markdown: /** @type {any} */ (markdown)
+		},
+		extends: ['markdown/recommended', 'markdown/processor']
+	},
 	{
 		name: 'local/language-options',
 		languageOptions: {
@@ -45,6 +54,7 @@ export default [
 		rules: {
 			'n/no-unsupported-features/es-builtins': 'error',
 			'n/no-unsupported-features/es-syntax': 'error',
+
 			'no-console': 'off',
 			'no-debugger': 'error',
 
@@ -93,8 +103,8 @@ export default [
 	{
 		name: 'local/inspector-extras',
 		files: [
-			'packages/vite-plugin-svelte-inspector/src/runtime/load-inspector.js',
-			'packages/vite-plugin-svelte-inspector/src/runtime/Inspector.svelte'
+			'packages/vite-plugin-svelte/src/plugins/inspector/runtime/load-inspector.js',
+			'packages/vite-plugin-svelte/src/plugins/inspector/runtime/Inspector.svelte'
 		],
 
 		languageOptions: {
@@ -137,7 +147,8 @@ export default [
 		files: ['**/*.md/*.js', '**/*.md/*.ts', '**/*.md/*.svelte'],
 		rules: {
 			'n/no-missing-import': 'off',
-			'@typescript-eslint/no-unused-vars': 'off'
+			'@typescript-eslint/no-unused-vars': 'off',
+			'no-undef': 'off'
 		}
 	},
 	{
@@ -170,4 +181,4 @@ export default [
 			'@typescript-eslint/no-unused-vars': 'off'
 		}
 	}
-];
+]);