+++ packages/e2e-tests/inspector-kit/svelte.config.js
@@ -2,10 +2,8 @@
 const config = {
 	kit: {},
 	vitePlugin: {
-		experimental: {
-			inspector: {
-				showToggleButton: 'always'
-			}
+		inspector: {
+			showToggleButton: 'always'
 		}
 	}
 };