+++ packages/vite-plugin-svelte-inspector/src/debug.js
@@ -1,3 +1,0 @@
-import _debug from 'debug';
-
-export const debug = _debug('vite-plugin-svelte-inspector');